:truck: Rename "Application Configuration" to "Metadata and Docs URLs"
to disambiguate between that and settings
