+++ docs/en/mkdocs.yml
@@ -65,7 +65,7 @@ nav:
   - tutorial/sql-databases.md
   - tutorial/bigger-applications.md
   - tutorial/background-tasks.md
-  - tutorial/application-configuration.md
+  - tutorial/metadata.md
   - tutorial/static-files.md
   - tutorial/testing.md
   - tutorial/debugging.md
